requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
cachetools
certifi
charset-normalizer
colorama
DateTime
distro
et_xmlfile
filetype
frozenlist
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
instaloader
jiter
jsonpatch
jsonpointer
langchain
langchain-core
langchain-google-genai
langchain-openai
langchain-text-splitters
langsmith
multidict
numpy
openai
opencv-python
openpyxl
orjson
packaging
pandas
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
pyee
pyparsing
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
requests-toolbelt
rsa
six
sniffio
SQLAlchemy
tenacity
tiktoken
tqdm
typing_extensions
tzdata
uritemplate
urllib3
yarl
yt-dlp
zope.interface
zstandard
